requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.11.7
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
supabase==2.3.4

# Vector Database
qdrant-client==1.15.1

# Graph Database
neo4j==5.16.0

# Document Processing
llama-parse==0.3.4
pypdf==3.17.4
python-docx==1.1.0
markdown==3.5.1
beautifulsoup4==4.12.2

# LLM & Embeddings
openai==1.6.1
google-generativeai==0.3.2
anthropic==0.8.1
sentence-transformers==2.2.2
tiktoken==0.5.2
scikit-learn==1.3.2
cohere==5.17.0
google-cloud-vision==3.7.0

# Async & Background Tasks
celery==5.3.4
redis==5.0.1
tenacity==8.2.3

# Source Connectors
notion-client==2.2.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
aiohttp==3.9.1
httpx==0.25.2
numpy==1.26.4
pandas==2.1.4
Pillow==10.1.0
python-multipart==0.0.6
